refactor: types 파일명 수정
