+++ src/recoil/Coupon/atoms/selectedCouponListState.ts
@@ -1,6 +1,6 @@
 import { atom } from 'recoil';
 
-import type { Coupon } from '../../../types/Coupon.type';
+import type { Coupon } from '../../../types/Coupon';
 
 export const selectedCouponListState = atom<Coupon[]>({
   key: 'selectedCouponListState',
