+++ src/recoil/DeliveryInfo/atoms/selectedDeliveryInfoListState.ts
@@ -1,6 +1,6 @@
 import { atom } from 'recoil';
 
-import type { DeliveryInfo } from '../../../types/DeliveryInfo.type';
+import type { DeliveryInfo } from '../../../types/DeliveryInfo';
 
 export const selectedDeliveryInfoListState = atom<DeliveryInfo[]>({
   key: 'selectedDeliveryInfoListState',
